Fix up unwanted partial file writes and implement event interface
The current runner system being based on Tower/AWX event processor
needs to handle events in a bit of a different way. In this case we'll
write the actual event data back to the filesystem and attach the
counter to the filename.

This also lets us build a nice generator interface to the events on
the Runner object at Runner.events