Fix an issue truncating the playbook stats lines
This issue likely exists in AWX/Tower also but is masked because the
events actually drive the stdout display entirely. In this case the
last status line omits the leading b64 encoded event structure and
this is not capture and emitted to stdout or the write buffer.

Here we force stdout to only contain what the event thinks the stdout
should be which is always going to be right and then account for the
situation where an event was emitted but no output needs to be displayed