fix up unit test errors for py3 support
This fixes an error where the test data was a string.  The test string
is now byte encoded to fully support py3.  This also fixes a number of
syntax errors to align with PEP8 rules

There was one change to method for generating the loader absolute path.
This change will now combine and expand the path if a relative path is
provided.  If a full path is provided, then the original value is
returned.  This simplifies the method to only a single return path.