merges envvars if one already exists for role execution
This change will now check the private_data_dir for an existing envvars
file and merge it with the roles path.  It will then restore the
original envvars file once the playbook run is done.

This change will also now generate a unqiue playbook name when executing
a role and remove the playbook once completed.

Finally this change will now remove the artifact lock files at the
conclusion if the playbook run instead of leaving them in the
private_data_dir.

The README file as been updated to include an example of how to use
ansible-runner to execute a role.  A test role has been added to the
demo folder.  To run the example execute the following command from the
project root folder.

`$ ansible-runner run ./demo -r testrole --roles-path ./demo/roles`