fixes bug that overwrites the inventory file
This fix addresses a bug where the inventory file will get overwritten
due to a default value for the hosts argument when running
ansible-runner from the command line.  This is due to a change that set
the default value for the hosts argument.  This fix removes the default
value and restores the operation to run as expected.