Enhance the logging and display output
This change provides a significant refactoring of the logging and
display facility to better provide output and debug functions.  It moves
the logging capabilities from utils into a separate singleton module
'output' and exposes methods to configure and produce output.

This change updates the entry points for both cli and api operation to
configuring the output.  To send output to there are two convience
functions available: `display` and `debug`.  The `display` function will
send output to `stdout` and to a log file (if enabled).  The `debug`
function will only produce output of the `--debug` option is specified
or the `debug=True` is set from the api interface.

Additionally all logging can be sent to a logfile using the `--logfile`
cli option or `logfile=<path>` for the api interface.