expose ansible-playbook verbosity to cli and api
This change adds a new config option to increase the verbosity of the
ansible playbook output.  In the command line form, you can now specify
`-v`, `-vv`, `-vvv`, etc and it will be passed to the ansible-playbook
at run time.  In the api form, set the kwarg `verbosity` to the number
of `-v`'s to append to `ansible-playbook`.