add new feature for mananging cmdline options
This change adds a new experimental feature to `ansible-runner` to
manage command line options passed to the `ansible-playbook` executable.
This feature does not validate the command line options so it is up to
the playbook writer to make sure to include the correct values.

In order to use the command line options create a new file in the `env/`
folder called `cmdline` and put the command line string into the file.

```
env/cmdline

--tags one,two --skip-tags three -u ansible --become
```

This change updates the docs as well