adds new option to skip fact collection when running a role directly
This change adds a new feature when executing a role directly that
allows the implementor to skip the fact collection.  By default facts
are collected but can be controlled by the option `--role-skip-facts`.
When the option is present on the command line when doing role
execution, `gather_facts` will be set to `False` in the playbook

implements #52