Don't show unnecessary tracebacks in the logs
There are a few parameters at runner config that are optional
but when such parameter is not used a traceback is shown in the logs.
This is a red herring in case one tries to debug the runner.

This patch puts the messages about missing file into a debug level.