Fix up timeout default values and update timeout handling
They were too low to be meaningful if the defaults were left in.

I've also updated the status text to reflect when a timeout has
occurred and picked a return code to match.

Lastly I've adjusted the idle timeout to be based on event receiving
and not based on the time since we last checked for a password since
this might not occur later on in long running playbooks