Fix extravar loader behavior
Our documentation makes it clear that extra vars passed in via the
programmatic interface are additive. In the presence of the
env/extravars file we should load that and then apply any extra vars
given in the parameterization afterwards.

Also, we were persisting some module-loaded data to disk which would
have the effect of writing the data out to files which is unexpected.

For a group of files (env/extravars, env/envvars, env/passwords,
env/settings, ssh_key, cmdline) if these files exist already then they
will not be overridden.