[utils.py] Fix currupt json stdout with large events
When using the `--json` output with `ansible-runner`, it is possible to
over fill `pexpect`'s `maxread` default of 2000, which ends up causing
the `else` of the `while` loop in the `OutputEventFilter.write` method
to get triggered, and that will indiscriminately push raw output to the
stdout handle.  This output is also a partial event at this point as
well, and shouldn't be pushed to STDOUT until a full event has been
processed and breaks the expectation of 1 line of JSON per line.

The intent of this section of the `else` is to provide some feedback to
the user if something goes wrong (for example, you provide a playbook
path that does not exist), as shown in the issue and pull request (fix)
where they originate from:

- https://github.com/ansible/ansible-runner/issues/29
- https://github.com/ansible/ansible-runner/pull/34

In addition, the changes from:

https://github.com/ansible/ansible-runner/pull/256

Are intended to make sure we treat individual lines that are not
currently parsed as events to be treated as unique "verbose events".

By combining the code that is used by #256 with the code from #34, we
can make sure that we wrap raw lines as events but not push content to
the STDOUT handler that shouldn't be there (partial events).