[test_display_callback] Use ANSIBLE_PYTHON_INTERPRETER var
Fixes test issues with new versions of ansible that have this patch:

  https://github.com/ansible/ansible/pull/50163

Since there are tests that end up triggering this warning and mess with
the event locations that are returned.

Made a quick refactor to the fixture in this test file as well to break
out the default env vars so it was a bit more DRY and re-usable.