simplifies the direct role execution path
This commit simplifies the role context manager and helps to stop
overloading variables.  Roles now execute in a temp directory only and
artifacts are not kept by default.  In order to keep artifiacts, the
`--artifacts-dir` command line option must be used.  This change also
enforces the use of `--hosts` instead of attaching to an existing
`private_data_dir`

Signed-off-by: Peter Sprygada <psprygad@redhat.com>