[test_display_callback.py] Add verbose output test
As part of the following PR description:

https://github.com/ansible/ansible-runner/pull/256

it was mentioned that there wasn't a integration test added for the
work, but a sample repo was provided here:

https://github.com/AlanCoding/ansible-runner-examples/tree/master/non_playbook/sleep_with_writes

with the manual steps used to replicate the desired behavior.  This
test is the codified version of that repo to ensure this functionality
exists after any refactorings take place.