adds new typing system to runner
This commit adds a new typing system to ansible-runner that allows
for programmatically buiding playbook and inventory objects.  The added
objects for playbook and inventory allow developers to construct a
playbook and/or inventory using native Python constructs instead of
having to create them out-of-band.  Once created the objects can be
serialized to JSON and injected directly to ansible-runner for
execution.

This commit also updates Pipefil with missing developer packages and
adds python coverage reporting to tox.

Signed-off-by: Peter Sprygada <psprygad@redhat.com>