Do not mangle roles_path when it is a string
Previously, when providing a single string as a parameter to `roles_path` the value would be mangled to insert colons between every character in the string. Now the assignment checks if the provided value is a single string or a list of strings and processes it into the environment variable accordingly.