Add ansible-runner-ubi8-container-image
Bring online ansible-runner devel image for ubi8. There are specific
things about this image, that will not work however. Because we need a
RHEL license to install 'subversion' we cannot do that.  It also means,
we wouldn't be able to redistrubute the image to quay.io too.

This is more to validate EEs could be built, using ansible-builder, then
actually functionally working.

Signed-off-by: Paul Belanger <pabelanger@redhat.com>