fix duplicate volume mounts for containers, simplify conditional
Uniformally handle dir paths that do have a trailing / and don't, but
are in fact the same path which are passed in or detected as duplicates
when constructing the list of volume mounts for containerized process
isolation.

Previously the BaseConfig._update_volume_mount_paths() method was using
this expression to check for a potentially duplicate path but the extra
processing of the map() and join() function calls are not needed since
the format src_mount_path:dest_mount_path stored in volume_mount_path is
already an element of arg_list and no transformation is necessary.

```
-        if ', '.join(map(str, ['-v', volume_mount_path])) not in ', '.join(map(str, args_list)):
+        if volume_mount_path not in args_list:
```

Signed-off-by: Adam Miller <admiller@redhat.com>
(cherry picked from commit 548a78fb31148ce0aa8fab6e63bd82bc8d37bbce)