Move utils/entrypoint.sh into ansible_runner/data folder
Because of poetry, there is no other way to do this if we want to
include the script in our python package.

We want to do this, so downstream container builds don't have to source
this file locally. We can use it from inside the container, is it will
now be installed.

Signed-off-by: Paul Belanger <pabelanger@redhat.com>