Do not rely on length of sys.argv
From the current behavior, it seems the goal is to print a common usage message
and exit 0 if no sub command was given. Under other circumstances, such as an
invalid or missing parameter, exit with 2 and print the normal usage message.

In order to preserve this behavior and make the tests run consistently whether
or not they are run in multiple process, a custom error method was created
that inspects the error message and exits 0 with the common usage message
if no command was given. This is a bit fragile, but inside that method
we do not have easy access to arguments passed to the main() function (they
could be inferred by inspecting private attributes, but this is probably
unwise).

This requires catching the SystemExit in the main() function and deciding there
based on sys_args how to handle it.

This fix preservers the existing behavior whil making the tests pass
consistently. However, it would be much simpler to rely on the default behavior
of argparse and exit 2 in both cases. This would remove the need for a custom
error method and exception handling. It would be a change in behavior, though.