Keep worker --private-data-dir when --delete not given (#887)
Keep worker --private-data-dir when --delete not given

I messed up our requirements a little bit with #820, and now I'm getting reports that people want the private_data_dir to be kept on the node running the ansible-runner worker command, and it is (unwantedly) getting deleted at the end of the run.
If you look at how I wrote the docs, I was dodgy about this case. I wrote it so that it would get cleaned up if it was a new directory. That didn't even work for the AWX use case, for which there is an intent to respect the setting AWX_CLEANUP_PATHS on both control and worker nodes.
I will take out of draft when I have confirmed this works with a production install.

Reviewed-by: David Shrewsbury <None>
Reviewed-by: None <None>
Reviewed-by: Sam Doran <sdoran@redhat.com>
Reviewed-by: Alan Rominger <arominge@redhat.com>