Add --ident option to process
This will allow users to write artifacts in a way that
matches normal non-remote ansible-runner commands

It makes the most sense to have the process --ident match
the --ident value passed to transmit, but you do not have to.

The intention is that this will be used by AWX so that
jobs ran and processed on the same node will not write
artifacts twice.