Write auto-created private data dirs to pytest dir instead of /tmp (#916)
Write auto-created private data dirs to pytest dir instead of /tmp

I briefly discussed this option with @samdoran in #901
After this patch, we create directories that look like:
/tmp/pytest-of-alancoding/pytest-0/popen-gw2/test_no_dst_all_dirs_src__dir_0/.ansible-runner-2fpvb4uo

This is as opposed to leaving around directories like /tmp/.ansible-runner-2fpvb4uo. There are a lot of these, like ~1,700 created with a single test run excluding the integration tests.
There are still a few directories created in /tmp after this, but it's a handful, not a flood.

Reviewed-by: Sam Doran <github@samdoran.com>
Reviewed-by: None <None>