Add option to disable streaming to stdout file (#937)
Add option to disable streaming to stdout file

This addresses #926
My current intention is to make use of the setting (set it to True) in AWX, unconditionally. We don't want the additional disk writes this incurs. If the executor team is okay with the looks of this, I may ask @amolgautam25 to make that PR.
This use case is only concerned with the pexpect runner mode. I only messed with subprocess here for completeness. The subprocess mode also appears to somehow be connected with the introduction of this behavior. Maybe @ganeshrn has more history to share about this.

Reviewed-by: John Westcott IV <None>
Reviewed-by: Alan Rominger <arominge@redhat.com>
Reviewed-by: David Shrewsbury <None>
Reviewed-by: None <None>