Preserve timestamps when unstreaming dirs (#966)
Preserve timestamps when unstreaming dirs

Connect #964
That fixes the linked issue with the limitation (discovered by the test) that we might miss a 2 second delta or less. This assumes the computer clock is perfectly synced on the machines that do the compressing and extracting.
Nonetheless, I think this opens up an option for AWX to get the job done. I show a demo in the test with the very_old.txt file, where I artificially set the mod and access time to 1 day in the past. It seems like that would work for the AWX use case, assuming that the two computers agree on the day, which doesn't sound like an unreasonable demand.
Ping @jbradberry

Reviewed-by: Jeff Bradberry <None>
Reviewed-by: David Shrewsbury <None>
Reviewed-by: Alan Rominger <arominge@redhat.com>
Reviewed-by: None <None>
(cherry picked from commit 203ef4314f00cc331c6f68d9c099f3707e78d891)