Mount the stdout callback plugin for containerized runs (#957)
Mount the stdout callback plugin for containerized runs

This is a continuation on from #763
I really want to keep @shanemcd's simplification which eliminates the aliasing to the minimal callback name. That PR was getting hung up on the fact that CI doesn't seem to build the new image correctly. This highlights a pain point that comes from the fact that ansible-runner used from the CLI employs the callback plugin from a different install of ansible-runner, baked into the image. Here, I'm investigating an option to address this pain point where I volume mount the stdout callback plugin, so we assure that any changes to the plugin will be instantly picked up without rebuilding. This might be overly-optimistic, but I'm hoping that this will be compatible with the old image, so even tests will pass. So far I've found some race conditions locally (probably due to multiprocessing), and depending on what tests give me here, my next step might be to make a per-run copy of the plugin.
So in short, instead of backing out some of our simplifications to get that other PR to merge, here, I'm entertaining the idea of leaning into even more simplifications that might hopefully solve several problems.

Reviewed-by: Alan Rominger <arominge@redhat.com>
Reviewed-by: Alexander Sowitzki <dev@eqrx.net>
Reviewed-by: David Shrewsbury <None>
Reviewed-by: Shane McDonald <me@shanemcd.com>
Reviewed-by: None <None>