Fix AAP-37599
ansible-runner will ignore adding the mountpoint to the '-v` argument list if
the source path does not exist. Instead it should respect whatever the
user is passing to the '--container-volume-mount' option. The exception would
be if the user tried to pass a file as the source path; ansible-runner should
resolve the file's parent directory.